requirements.txt
Django>=3.0.7
psycopg2
djangorestframework>=3.9.1
pylint
pydash
django-cors-headers
pyyaml
colour_runner
boto3
moto
git+https://github.com/snyaggarwal/django-queryset-csv
six
python-dateutil
requests
factory_boy
